libblkid: fix 'partno' usage
Linux kernel ignores empty partitions, but partition number is always
allocated for the partition. (This rule is used for primary partitions
only.)

For example:

  part 1.  size=12345
  part 2.  size=0
  part 3.  size=24567

the final list of partitions:

  sda1 sda3

'sda2' is not defined and partno=2 is not used for any other
partitions.

The libblkid library has to be compatible with this rule.

Signed-off-by: Karel Zak <kzak@redhat.com>