fdisk: improve 'move begin of partition' command
The 'b' command ("move beginning of data in a partition")
implementation is too restricted. For example:

     Device Boot      Start         End      Blocks   Id  System
  /dev/sdb1            2048       22527       10240   83  Linux
  ....
  Partition number (1-4): 1
  New beginning of data (2048-22527, default 2048):
                         ^^^^^^^^^^

the range is defined by the current partition size. New version:

  Partition number (1-4): 1
  New beginning of data (1-22527, default 2048):
                         ^^^^^^^

allows to move the begin to arbitrary place if the place is not
allocated by any other partition. This is useful for people who don't
want to use the default partitioning (1MiB offset, 1MiB grain).

The 'b' command is expert command and does not force users to use an
aligned LBA (fdisk(8) still prints warning (in 'p' command) if any
partition is not aligned to the physical sector boundary).

Signed-off-by: Karel Zak <kzak@redhat.com>