fdisk: extend -c and -u options to support old DOS stuff
The DOS mode and cylinders as display units are disabled by default
now. For users who need old DOS behavior is it necessary to extend
-c (compatibility) and -u (units) command line options. The change is
backwardly compatible (so people who already uses -u and -c are not
affected with this change). New form:

 -c[=<mode>]   where <mode> is 'dos' or 'nondos' (default)
 -u[=<units>]  where <units> is 'cylinders' or 'sectors' (default)

For example old deprecated DOS behavior could be enabled by:

  fdisk -u=cylinders -c=dos /dev/sda

Signed-off-by: Karel Zak <kzak@redhat.com>