swapon: Document btrfs limitation with swapfiles
Btrfs, as of 2.6.35, is unable to allow swapfiles to be used on its
filesystems. This is due to the swapfile implementation wanting to
build an extent map of each block in the file and expecting it to be
static for the life of the swapfile.

Btrfs can't guarantee this and refuses to return the mapping. The
swapfile implementation just makes a comment about there being holes
in the file - but that's how btrfs denies the mapping.

This patch adds a section to the swapon manpage to document it.

Signed-off-by: Jeff Mahoney <jeffm@suse.com>
Signed-off-by: Karel Zak <kzak@redhat.com>