flock: fix file number in flock.1 man page
According to a search on the internet, the 200 file number is not a
posix shell feature, since posix shells are only mandated to have 1
through 9.  I haven't verified that this claim is right, but, at least
dash doesn't like the 200. So here is a patch for the flock man page
that replaces 200 with 9.

Signed-off-by: Patrice Dumas <pertusus@free.fr>
Signed-off-by: Karel Zak <kzak@redhat.com>