findmnt: add support for fs-root (subvolumes and bind mounts)
This patch modifies the default output for SOURCE column. All btrfs
subvolume mountpoints and all bind-mount (where source is not root of
FS) will be printed as:

   SOURCE               TARGET
   /dev/sda1[/aaa]      /mnt/test

where /aaa is subvolume name or fs root for bind mounts, it means:

   # mount -t btrfs /dev/sda1 /mnt/test -o subvol=aaa

or:

   # mount --bind /aaa /mnt/test

The info about fs-root is 4th column in /proc/self/mountinfo.

Signed-off-by: Karel Zak <kzak@redhat.com>