more: new usage output, and some trivial fixes
Usage output screen got to be completely rewrote. Aim is to
assist user to determe which switches to use without reading
manual page. The change also introduced new option to see command
version.

Determination if the executable name is `page' is now easy to see
in source, and effect of using the name is mentioned in manual
page.

Rest of the change is trivial; exit values to compiler warning
removals, magic constants to preprocessor definitions, symbolic
exit values, comment pretty printing and #endif markups.

[kzak@redhat.com: - coding style
                  - add __unused__ attribute to signal handlers]

Signed-off-by: Sami Kerola <kerolasa@iki.fi>
Signed-off-by: Karel Zak <kzak@redhat.com>