umount: allow unmounting loopdev specified by associated file
Make it possible to unmount a filesystem on a loop device if it is
specified by associated backing file. It does not attempt to unmount
anything if there are more than one loop device associated with the
given file.

Umount looks for associated loopdevice(s) only if umount is called
with the regular file as an argument.

Before:
	mount -o loop -t ext2 img mnt
	umount -v img
	> Could not find /home/puzel/upstream/util-linux/img in mtab
	> umount: img: not mounted

After:
	mount -o loop -t ext2 img mnt
	umount -v img
	> img is associated with /dev/loop0, trying to unmount it
	> /dev/loop0 has been unmounted

[kzak@redhat.com: - fix memory leak in lomount.c]

Addresses: https://bugzilla.novell.com/show_bug.cgi?id=666161
Signed-off-by: Petr Uzel <petr.uzel@suse.cz>
Signed-off-by: Karel Zak <kzak@redhat.com>