libmount: add support for x-* mount comments
The current way how add any comment or extra information to /etc/fstab
is to use comment="" mount option. This is not too elegant and readable
if you want to add more than one comment.

It seems better to add any generic prefix that will be used for all
3rd-party userspace mount options.

Expected syntax is:

  x-<application>-<options>[=<data>]

For example:

  x-systemd-bar=managed

All x-* options (as well as comment=) are ignored by libmount and not
stored to the mtab file.

Requested-by: Lennart Poettering <lennart@poettering.net>
Signed-off-by: Karel Zak <kzak@redhat.com>