docs: rename.1 verbose, long options and warning
Add new verbose and long options to manual page. I also added
warning section which hopefully gets read by people who run the
command first time.

It might be good idea to make the rename such that it would not
overwrite by default, and have a --force option if an use wants
that. In current state the rename is somewhat dangerous.

Signed-off-by: Sami Kerola <kerolasa@iki.fi>