mount: first look for mountpoint
  # mount <device|dir>

The current code looks for a device and then for a mountpoint in
/etc/fstab. This is not user friendly solution. People usually use

 # mount /dir

to mount any filesystem. It makes more sense to check for mountpoint
and if not found then for device.

This is also important for bind mounts, for example if you have in
your fstab:

	/dev/sda1	/mnt/foo	auto defaults
	/mnt/foo	/mnt/bar	none bind

then
	# mount /mnt/foo

should be interpreted as the first entry and /dev/sda1 should be
mounted.

Addresses: https://bugzilla.redhat.com/show_bug.cgi?id=716483
Signed-off-by: Karel Zak <kzak@redhat.com>