mkfs.minix: add minix v3 support
We can now create minix v3 filesystems. Support for this fs was added a few
years ago in the Linux kernel. One of the most important benefits is the
ability to handle file names up to 60 characters long.  With this change we
also introduce the -3 option which naturally indicates which version to create.
Version 1 is still left as the default one for backwards compatibility reasons.

[kzak@redhat.com: - fix gcc warnings:

	mkfs.minix.c: In function ‘super_set_state_ptr’:
	mkfs.minix.c:174:1: warning: no return statement in function returning non-void [-Wreturn-type]
	mkfs.minix.c: In function ‘write_tables’:
	mkfs.minix.c:182:7: warning: unused variable ‘state’ [-Wunused-variable]
	mkfs.minix.c: At top level:
	mkfs.minix.c:111:27: warning: ‘blksz’ defined but not used [-Wunused-variable]
]

Signed-off-by: Davidlohr Bueso <dave@gnu.org>
Signed-off-by: Karel Zak <kzak@redhat.com>