mount: don't canonicalize the source of a pseudofs
In line with previous patches, the source of a pseudofs is strictly
semantic; do not attempt to resolve it to a filesystem path.

As a side effect of this, nspec needs to be initialized to NULL since
it may not be assigned null by spec_to_devname when the FS is deemed
a pseudofs.

Signed-off-by: Dave Reisner <dreisner@archlinux.org>